Create ComplexNumber.class
Source code for the immutable class ComplexNumber, just basic stuff and to be updated over the course of being used until it feels like it has everything it needs. I'm not particularly interested in supporting things like sin(z) but we'll see. I'm more interested in using this for physics as well as eventually using it as elements of a matrix class maybe.