Uploaded incomplete p-adic DFT matrix
It only finds the first digit of the p-adic numbers, but it works as far as finding the inverses. Combined with the hensel_finder.m it will be able to create (p-1)x(p-1) invertible matrices of p-adic numbers, which really will appear to be real valued matrices doing Fourier transforms since the field of p-adics contains all the p-1 roots of unity.